requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
Pillow==6.1.0
pkg-resources==0.0.0
slackclient==2.1.0
typing-extensions==3.7.4
yarl==1.3.0
